+++ cantina/src/main/java/com/igrejacristahangar/cantina/modules/pedido/dto/PedidoResumo.java
@@ -3,9 +3,11 @@ package com.igrejacristahangar.cantina.modules.pedido.dto;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 public class PedidoResumo {
-    private Double totalPreco;
-    private Double totalPrecoDeHoje;
+    private BigDecimal totalPreco;
+    private BigDecimal totalPrecoDeHoje;
  }
